papergrid: add full support of row span
It changes a lot of code, and makes it overly not outstanding.
Indeed it deserve refactoring.

It addes `block` logic, which split rows by blocks with the same pattern.
And since we dismantle rows we should tune blocks as well as we did to rows.
I mean we should basically remove the bottom border of
each block not included the last one.
This changes calculates relational weight for each block to adjust them.
It basically happen in the rows 181-206.
This must be done since sometimes block not capable to set its weight to a specific one.

There is a critical issue is we remove display of `cell.ident.left` and `cell.ident.right`
from CellFormater. Which is seems a logical discrepancy.
Not it's done it grid formatting logic by calculating
weight = c.weight + cell.ident.left + cell.ident.right (look at line 134).
From my perspective it shows us style of problems in the design.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>