papergrid: restructure code
Add a new unpromted level of abstraction `portions`. Portions is
responsible for spliting the grid in proper ratio.
And put back responsibility of showing indents to `CellFormatter`.

It brings a modification of basic style. And what's more
it affects a cell relation logic. You can see on test
`render_row_span_with_horizontal_ident` to reach it.

Basically before all the work was doing in fmt function.
Which is not good? when you are going through it's
get bored almoust emidiately. Personally I was ambush lots of time
pondering what should be changed to add new feature or change something.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>