papergrid: add support column span
It involves rewriting algorithm of size detirmination and
printing algorithm.

Updates the test suite with new test cases.

This PR describes main idea. But the code is amusingly not good state.
And what is more important there's still a pile of pitfalls.
It's linked to row/column indents and style in cell's borders.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>