tabled/ Rename Join to Concat
The current implementation of join doesn't respond,
to sql style of join.
But it does have a tyle of concat of operation.
Where we just concat 2 tables together and that's it.

I am not sure even though if current implementation is finished.
Because looking at python pandas they have some stuff like indexes etc.
I hope that this type of abstractions only needed in sql join.

Signed-off-by: Maxim Zhiburt <zhiburt@gmail.com>