tabled/ Fix IndexBuilder on empty tables (#363)
There's a minor behaviour difference between how `Builder` and `IndexBuilder`
handles tables that have no data rows but a header row. In that case,
`Builder` will regard the header row itself as data, but `IndexBuilder`
will not. This difference shows up when `table.is_empty()` is called
on the resulting tables: such tables will have different shapes.

This patch fixes the process when an `IndexBuilder` is converted back
to a `Builder`. With this patch, a default `Builder` will be returned
only if the `IndexBuilder` has neither data rows nor a header row.

Unit tests are added accordingly.